// my-courses-lab4-beek1800/src/app/models/MyCourse.ts
export class MyCourse{
    completed: boolean = false;
    courseCode: string = "courseCode";
    subjectCode: string = "subjectCode";
    level: string = "level";
    name: string = "name";
    points: number = 0;
    institutionCode: string = "institutionCode";
    subject: object = { subject: "subject" };
}